requirements.txt
python-telegram-bot==20.3
aiohttp==3.8.5
sqlite3-binary==0.0.1